# PERF: Allow str.split callers to skip expensive post-processing