@jreback 2020-07-11
we already have some benchmarks for this, so should be easy to tell a good change.