@simonjayhawkins 2020-07-24
@wbadart can you address comments?