@w-m 2021-08-30
This [stackoverflow answer](https://stackoverflow.com/a/68990278/463796) highlights a performance issue that is likely related. In the example data, there are rows of strings separated by commas with a varying number of elements. str.split with `expand=True` is about 2x slower than str.split with `expand=False` followed by `to_list()` and creating a new DataFrame from the results:

```
df = pd.DataFrame(["a", "a,b", "a,b,c"] * 100000, columns=["steps"])

%timeit res = df.steps.str.split(",", expand=True)
249 ms ± 12.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

%timeit res2 = pd.DataFrame(df.steps.str.split(",").to_list())
128 ms ± 1.31 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)


```